Refactored integration tests and moved them to a dedicated package
A superclass was introduced that the integration tests now extend, which allows configuration and utility methods to be shared.
