+++ crown-court-proceeding/src/test/java/uk/gov/justice/laa/crime/crowncourt/integration/MagsProceedingIntegrationTest.java
@@ -1,51 +1,27 @@
-package uk.gov.justice.laa.crime.crowncourt.controller;
+package uk.gov.justice.laa.crime.crowncourt.integration;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.github.tomakehurst.wiremock.WireMockServer;
 import com.github.tomakehurst.wiremock.client.WireMock;
-import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.TestInstance;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.actuate.observability.AutoConfigureObservability;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.cloud.contract.wiremock.AutoConfigureWireMock;
-import org.springframework.context.annotation.Import;
 import org.springframework.http.HttpMethod;
 import org.springframework.http.MediaType;
 import org.springframework.security.web.FilterChainProxy;
-import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 import org.springframework.web.context.WebApplicationContext;
-import uk.gov.justice.laa.crime.crowncourt.config.IntegrationTestConfiguration;
 import uk.gov.justice.laa.crime.crowncourt.data.builder.TestModelDataBuilder;
 import uk.gov.justice.laa.crime.crowncourt.dto.maatcourtdata.RepOrderDTO;
-import uk.gov.justice.laa.crime.crowncourt.model.common.ApiFinancialAssessment;
-import uk.gov.justice.laa.crime.crowncourt.model.common.ApiIOJSummary;
-import uk.gov.justice.laa.crime.crowncourt.model.request.ApiDetermineMagsRepDecisionRequest;
-import uk.gov.justice.laa.crime.enums.CaseType;
 import uk.gov.justice.laa.crime.enums.DecisionReason;
 import uk.gov.justice.laa.crime.util.RequestBuilderUtils;
 
 import java.time.LocalDate;
-import java.util.Map;
-import java.util.UUID;
 
-import static com.github.tomakehurst.wiremock.client.WireMock.post;
 import static com.github.tomakehurst.wiremock.client.WireMock.put;
-import static org.springframework.boot.test.context.SpringBootTest.WebEnvironment.DEFINED_PORT;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
 
-@DirtiesContext
-@AutoConfigureObservability
-@AutoConfigureWireMock(port = 9998)
-@Import(IntegrationTestConfiguration.class)
-@TestInstance(TestInstance.Lifecycle.PER_CLASS)
-@SpringBootTest(classes = IntegrationTestConfiguration.class, webEnvironment = DEFINED_PORT)
-class MagsProceedingIntegrationTest {
+class MagsProceedingIntegrationTest extends WiremockIntegrationTest {
 
     private MockMvc mvc;
 
@@ -58,15 +34,7 @@ class MagsProceedingIntegrationTest {
     @Autowired
     private WebApplicationContext webApplicationContext;
 
-    @Autowired
-    private WireMockServer wiremock;
-
     private static final String ENDPOINT_URL = "/api/internal/v1/proceedings/determine-mags-rep-decision";
-
-    @AfterEach
-    void clean() {
-        wiremock.resetAll();
-    }
 
     @BeforeEach
     public void setup() {
@@ -89,13 +57,7 @@ class MagsProceedingIntegrationTest {
     @Test
     void givenInvalidRequest_whenDetermineMagsRepDecisionIsInvoked_thenFailsBadRequest() throws Exception {
         var apiUpdateApplicationRequest =
-                new ApiDetermineMagsRepDecisionRequest()
-                        .withRepId(null)
-                        .withCaseType(CaseType.INDICTABLE)
-                        .withUserSession(TestModelDataBuilder.getApiUserSession(true))
-                        .withIojAppeal(new ApiIOJSummary().withIojResult("PASS"))
-                        .withPassportAssessment(TestModelDataBuilder.getPassportAssessment())
-                        .withFinancialAssessment(new ApiFinancialAssessment().withInitResult("PASS"));
+                TestModelDataBuilder.getApiDetermineMagsRepDecisionRequest(false);
 
         var applicationRequestJson = objectMapper.writeValueAsString(apiUpdateApplicationRequest);
         mvc.perform(RequestBuilderUtils.buildRequestGivenContent(HttpMethod.POST, applicationRequestJson, ENDPOINT_URL))
@@ -105,13 +67,7 @@ class MagsProceedingIntegrationTest {
     @Test
     void givenAPIClientException_whenDetermineMagsRepDecisionIsInvoked_thenFailsServerError() throws Exception {
         var apiUpdateApplicationRequest =
-                new ApiDetermineMagsRepDecisionRequest()
-                        .withCaseType(CaseType.INDICTABLE)
-                        .withRepId(TestModelDataBuilder.TEST_REP_ID)
-                        .withUserSession(TestModelDataBuilder.getApiUserSession(true))
-                        .withIojAppeal(new ApiIOJSummary().withIojResult("PASS"))
-                        .withPassportAssessment(TestModelDataBuilder.getPassportAssessment())
-                        .withFinancialAssessment(new ApiFinancialAssessment().withInitResult("PASS"));
+                TestModelDataBuilder.getApiDetermineMagsRepDecisionRequest(true);
 
         stubForOAuth();
         wiremock.stubFor(put("/api/internal/v1/assessment/rep-orders")
@@ -130,13 +86,7 @@ class MagsProceedingIntegrationTest {
     @Test
     void givenValidRequest_whenDetermineMagsRepDecisionIsInvoked_thenSucceeds() throws Exception {
         var apiUpdateApplicationRequest =
-                new ApiDetermineMagsRepDecisionRequest()
-                        .withCaseType(CaseType.INDICTABLE)
-                        .withRepId(TestModelDataBuilder.TEST_REP_ID)
-                        .withUserSession(TestModelDataBuilder.getApiUserSession(true))
-                        .withIojAppeal(new ApiIOJSummary().withIojResult("PASS"))
-                        .withPassportAssessment(TestModelDataBuilder.getPassportAssessment())
-                        .withFinancialAssessment(new ApiFinancialAssessment().withInitResult("PASS"));
+                TestModelDataBuilder.getApiDetermineMagsRepDecisionRequest(true);
 
         stubForOAuth();
         wiremock.stubFor(put("/api/internal/v1/assessment/rep-orders")
@@ -159,22 +109,5 @@ class MagsProceedingIntegrationTest {
                 .andExpect(jsonPath("$.decisionResult.decisionReason").value(DecisionReason.GRANTED.getCode()))
                 .andExpect(jsonPath("$.decisionResult.timestamp").value(
                         TestModelDataBuilder.TEST_DATE_MODIFIED.toString()));
-    }
-
-    private void stubForOAuth() throws JsonProcessingException {
-        ObjectMapper mapper = new ObjectMapper();
-        Map<String, Object> token = Map.of(
-                "expires_in", 3600,
-                "token_type", "Bearer",
-                "access_token", UUID.randomUUID()
-        );
-
-        wiremock.stubFor(
-                post("/oauth2/token").willReturn(
-                        WireMock.ok()
-                                .withHeader("Content-Type", String.valueOf(MediaType.APPLICATION_JSON))
-                                .withBody(mapper.writeValueAsString(token))
-                )
-        );
     }
 }
